requirements.txt
asgiref==3.7.2
chardet==5.2.0
Django==4.2.10
django-extensions==3.2.3
django-filter==24.1
django-tables2==2.7.0
et-xmlfile==1.1.0
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
pillow==10.2.0
psycopg==3.1.18
psycopg-binary==3.1.18
python-dateutil==2.9.0.post0
pytz==2024.1
reportlab==4.1.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
